% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{est}
\alias{est}
\title{Estimates and SEs from a sample}
\usage{
est(s, pop, fmla, ...)
}
\arguments{
\item{s}{a sample data frame}

\item{pop}{a population data frame}

\item{fmla}{a formula of the form \code{population ~ a + b + c} giving a variable with population 
counts in \code{pop} and stratification variables.}

\item{\dots}{other arguments passed to \code{\link{HTwts}}}
}
\value{
a list with parameter estimates and standard errors
}
\description{
Use \code{\link{ns}}, \code{\link{HTwts}} and other functions to compute estimated parameters
and SEs with different algorithms
}

