% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{HTwts}
\alias{HTwts}
\title{Horvitz-Thompson and related weights}
\usage{
HTwts(sample, cap = 3, ...)
}
\arguments{
\item{sample}{a sample data frame with sample and population stratum counts given by the
variables 'n' and 'N' respectively.}

\item{cap}{(default 3) a maximum for truncated median normed Horvitz-Thompson weights.}

\item{\dots}{other arguments for future weight functions}
}
\description{
Calculates sample weights based on the sample and population count variables calculated by
\code{\link{ns}}.
}

