% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/survey.R
\name{popagetable}
\alias{popagetable}
\title{Population table (IN PROGRESS)}
\format{A data frame with 450800 rows and 8 variables.
\describe{
  \item{country}{a factor: US or PR}
  \item{state}{a factor: US state, DC or PR}
  \item{state.code}{2-letter state abbreviation and 2-digit code containing the first 2 digits
  of county codes}
  \item{sex}{a factor: Male or Female}
  \item{age}{an ordered factor with 14 age groups}
  \item{raceethnicity}{a factor with 4 categories plus 'OTHER'}
  \item{population}{a numeric variable with the population (IN PROGRESS -- USE ONLY FOR TESTING) in
  each row. Negative values in 'acsagetable' have been set to zero}
  \item{population_neg}{a numeric variable with values of population in 'acsagetable' including
  negative values}
}}
\source{
'acsagetable' from the WWC package. See the script creating this data frame in 'etc/popagetable-data.R'

Datassist plus some arbitraty adjustments for population
}
\description{
Population table by sex, age group (14 groups), raceethnicity (5 categories) within country (US and Puerto Rico),
state (50 states plus DC and PR) and county.

This data frame should be used only for testing
}
\details{
This data frame contains the number of individuals within each county in the 
US and Puerto Rico (PR) in clusters formed by combining categories of 
age (in 14 categories), raceethnicity (in 5 categories) and sex (Male or Female).
Each county is also identified by its 5-digit county code,  
its state (50 states using 2-letter abbreviations plus DC and PR),
and its country (US or PR).
}

