% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{sam}
\alias{sam}
\title{Sample from a sampling frame}
\usage{
sam(N = 1, pop = popagetable, fmla = population ~ sex + age +
  raceethnicity + state, subset, fac, prob)
}
\arguments{
\item{N}{sample size (default: 1)}

\item{pop}{a data frame giving frequencies in each stratum defined by a combination of
variables. The stratification can be finer than the stratification given in \code{fmla} (Default: \code{popagetable})}

\item{fmla}{a formula of the form \code{count ~ a + ... + c} where \code{count} is a variable in
\code{pop} giving the population count in each row of \code{pop} and the right hand side 
of the formula identifies the variables used to aggregate \code{pop} to form a sampling frame
(Default: \code{population ~ sex + age + raceethnicity + state}). The default aggregates over
counties in the default value of \code{pop}}

\item{subset}{an expression that selects a subset, e.g. state == 'AK' & age > '5 to 9 years' (Default: no subsetting)}

\item{fac}{a data frame or a list of data frames whose variables include some subset of variables in the
right hand side of \code{fmla} and a variable named \code{fac} giving relative sampling factors
for each combination of values of the stratification variables.  Omitted combinations of variables get
a default relative sampling factor of 1. If \code{fac} is a list, the relative sampling factors
from each element are multiplied.}

\item{prob}{a data frame or a list of data frames similar to \code{fac}, except that a variable named
\code{prob} is used to specify the binomial probability of a '1' in each stratum of respondents.
If \code{prob} is a list, the probabilities from each data frame are combined by adding their
logits.}
}
\description{
Generate a probability sample from a sampling frame
  
\code{sam} takes a sample size, a 'population' data frame, a formula specifying the count variable and 
stratification variable(s), an optional expression to select a subset of the 
population, relative sampling factor and the probability parameter for
a binomial response. It generates a sample based on these parameters.
}

