% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{wtd_mean}
\alias{wtd_mean}
\title{Weighted mean}
\usage{
wtd_mean(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{numerical vector}

\item{w}{vector of weights replicated to have same length as \code{x}}

\item{na.rm}{if TRUE, drop units for which x is NA. Default: FALSE.}
}
\description{
Weighted mean
}
\seealso{
\code{\link{wtd_mean}} for weighted means, \code{link{lin_comb}} for linear
     combinations, \code{\link{jk_wtd_mean_se}} for a jacknife estimate of the SE of a weighted mean
     and \code{\link{jk_lin_comb_se}} for a jacknife estimate of the SE of a linear combination.

Other Post-stratification survey functions: \code{\link{jk_wtd_mean_se}},
  \code{\link{jk_wtd_means}}, \code{\link{lin_comb}}
}

